Install disaggregator via pip in seperate task
